dom/indexedDB/test/test_create_objectStore.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const nsIIDBObjectStore = Components.interfaces.nsIIDBObjectStore;
      const nsIIDBTransaction = Components.interfaces.nsIIDBTransaction;

      const name = window.location.pathname;
      const description = "My Test Database";
      const objectStoreInfo = [
        { name: "1", options: { keyPath: "" } },
        { name: "2", options: { keyPath: "", autoIncrement: true } },
        { name: "3", options: { keyPath: "", autoIncrement: false } },
        { name: "4", options: { keyPath: "" } },
        { name: "5", options: { keyPath: "foo" } },
        { name: "6" },
        { name: "7", options: null },
        { name: "8", options: { autoIncrement: true } },
        { name: "9", options: { autoIncrement: false } },
        { name: "10", options: { keyPath: "foo", autoIncrement: false } },
        { name: "11", options: { keyPath: "foo", autoIncrement: true } }
      ];

      let request = mozIndexedDB.open(name, description);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      ok(event.target.source === mozIndexedDB,
         "event.target.source is correct");
      let db = event.target.result;

      let count = db.objectStoreNames.length;
      is(count, 0, "correct objectStoreNames length");

      request = db.setVersion("1");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      ok(event.target.source === db, "event.target.source is correct");

      try {
        db.createObjectStore(null);
        ok(false, "createObjectStore with null name should throw");
      }
      catch(e) {
        ok(true, "createObjectStore with null name should throw");
      }

      try {
        db.createObjectStore("");
        ok(false, "createObjectStore with empty name should throw");
      }
      catch(e) {
        ok(true, "createObjectStore with empty name should throw");
      }

      try {
        db.createObjectStore("foo", "bar");
        ok(false, "createObjectStore with bad options should throw");
      }
      catch(e) {
        ok(true, "createObjectStore with bad options");
      }

      try {
        db.createObjectStore("foo", { foo: "" });
        ok(false, "createObjectStore with bad options should throw");
      }
      catch(e) {
        ok(true, "createObjectStore with bad options");
      }

      for (let index in objectStoreInfo) {
        index = parseInt(index);
        const info = objectStoreInfo[index];

        let objectStore = info.hasOwnProperty("options") ?
                          db.createObjectStore(info.name, info.options) :
                          db.createObjectStore(info.name);

        is(db.objectStoreNames.length, index + 1,
           "updated objectStoreNames list");

        let found = false;
        for (let i = 0; i <= index; i++) {
          if (db.objectStoreNames.item(i) == info.name) {
            found = true;
            break;
          }
        }
        is(found, true, "objectStoreNames contains name");

        is(objectStore.name, info.name, "Bad name");
        is(objectStore.keyPath, info.options && info.options.keyPath ?
                                info.options.keyPath : "",
           "Bad keyPath");
        if(objectStore.indexNames.length, 0, "Bad indexNames");

        ok(event.target.transaction, "event has a transaction");
        ok(event.target.transaction.db === db, "transaction has the right db");
        is(event.target.transaction.readyState, nsIIDBTransaction.LOADING,
           "transaction has the correct readyState");
        is(event.target.transaction.mode, nsIIDBTransaction.VERSION_CHANGE,
           "transaction has the correct mode");
        is(event.target.transaction.objectStoreNames.length, index + 1,
           "transaction has correct objectStoreNames list");
        found = false;
        for (let j = 0; j < event.target.transaction.objectStoreNames.length;
             j++) {
          if (event.target.transaction.objectStoreNames.item(j) == info.name) {
            found = true;
            break;
          }
        }
        is(found, true, "transaction has correct objectStoreNames list");
      }

      finishTest();
      yield;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>
</head>

<body onload="runTest();"></body>

</html>
